[MOVE/#6] 파일 이동 및 분리 - core/designsystem으로 테마 파일 이동
