+++ app/src/main/java/com/sopt/shinmungo/core/designsystem/color/Color.kt
@@ -1,0 +1,35 @@
+package com.sopt.shinmungo.core.designsystem.color
+
+import androidx.compose.runtime.Immutable
+import androidx.compose.runtime.staticCompositionLocalOf
+import androidx.compose.ui.graphics.Color
+
+//Primary Colors
+val Orange = Color(0xFFFE6A00)
+val Beige = Color(0xFFFFB987)
+val Red = Color(0xFFD60000)
+
+//Grey Scale Colors
+val Gray1 = Color(0xFFFFFFFF)
+val Gray2 = Color(0xFFFAFAFA)
+val Gray3 = Color(0xFFF1F1F1)
+val Gray4 = Color(0xFFE7E7E7)
+val Gray5 = Color(0xFFCFCFCF)
+val Gray6 = Color(0xFFA4A4A4)
+val Gray7 = Color(0xFF818181)
+val Gray8 = Color(0xFF717171)
+val Gray9 = Color(0xFF656565)
+val Gray10 = Color(0xFF545454)
+val Gray11 = Color(0xFF3F3F3F)
+val Gray12 = Color(0xFF282828)
+val Gray13 = Color(0xFF121212)
+
+//Opacity
+val OpacityGray13_60 = Color(0x99121212)
+val OpacityGray13_40 = Color(0x66121212)
+val OpacityGray13_10 = Color(0x1A121212)
+val OpacityGray13_5 = Color(0x0D121212)
+val OpacityGray1_30 = Color(0x4DFFFFFF)
+val OpacityGray1_10 = Color(0x1AFFFFFF)
+
+
