+++ app/src/main/java/com/sopt/shinmungo/core/designsystem/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.sopt.shinmungo.ui.theme
+package com.sopt.shinmungo.core.designsystem.theme
 
 import android.app.Activity
 import androidx.compose.material3.MaterialTheme
@@ -10,6 +10,10 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalView
 import androidx.core.view.WindowCompat
+import com.sopt.shinmungo.core.designsystem.color.Gray1
+import com.sopt.shinmungo.core.designsystem.color.LocalColorProvider
+import com.sopt.shinmungo.core.designsystem.color.ShinMunGoColors
+import com.sopt.shinmungo.core.designsystem.color.defaultShinMunGoColors
 
 object ShinMunGoTheme {
         val color: ShinMunGoColors
